[!!!][TASK] Remove tab handlers and improve code style
Since the release of TYPO3 6, large portions of backend code have
been improved.
This change completely removes support for so called "tab handlers"
from the linkhandler extension to make the adoption of latest changes
in the TYPO3 backend code possible.
In this context, code that integrates the linkhandler extension in the
TYPO3 backend has been refactored.